Updated to latest Aviary version & changed from using events to callbacks (which seem to be more reliable)
Updated Aviary libray to latest version.
Added Foundation.CoreText (which is required for the latest version) to
the module.xml file.
Changed the way the image or cancelation is passed back to the caller
to use a function callback as opposed to event handler (which wasnt
working for some reason).